Modified-U-Net-Change-Detection/requirements.txt
streamlit
numpy==1.26.4
torch==2.3.0
torchvision
Pillow
opencv-python-headless
matplotlib
albumentations
imgaug
pandas
plotly
scikit-learn
fpdf2
google-generativeai
googletrans
streamlit-image-comparison
streamlit-drawable-canvas
python-dotenv
nest-asyncio
pyyaml
tqdm
scipy
